% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coGSEA.r
\name{coGSEA}
\alias{coGSEA}
\title{Wrapper function of coGSEA}
\usage{
coGSEA(ElistObject, contrastMatrix, ENTREZGenesIds, geneSetCollection = "H",
  specie = "Mus musculus", directoryPath = "./", alpha = 0.05,
  pvalAdjMethod = "BH", pvalCombMethod = "sumlog",
  min.intersection.size = 1, GSEA.Methods = c("camera", "gage",
  "globaltest", "gsva", "ssgsea", "zscore", "plage", "ora", "padog", "roast",
  "safe"), num.workers = 4, shinyMode = FALSE)
}
\arguments{
\item{ElistObject}{An Elist object.}

\item{contrastMatrix}{A contrast matrix.}

\item{ENTREZGenesIds}{vector of gene ENTREZ identifiers.}

\item{geneSetCollection}{H, C2_KEGG, or C2_REACTOME, default to H (character). Can be custom of class list.}

\item{specie}{Mus musculus or Homo sapiens, default to Mus Musculus (character).}

\item{directoryPath}{path to existing directory where to save the results ("character").}

\item{alpha}{alpha treshold, default to 0.05 (numeric).}

\item{pvalAdjMethod}{p.value adjustment method (character).}

\item{pvalCombMethod}{p.value combination method (character).}

\item{min.intersection.size}{minimum intersection size for SnailPlot (integer)}

\item{GSEA.Methods}{GSEA methods to run (vector)}

\item{num.workers}{number of thread to use (integer)}

\item{shinyMode}{Make plot or not (boolean)}
}
\value{
A result list if shinyMode == TRUE, else nothing
}
\description{
Wrapper function of coGSEA
}
\examples{
\dontrun{
coGSEA(ElistObject = Elist, contrastMatrix = contrast, ENTREZGenesIds = Elist$genes$ENTREZID, directoryPath = "./")
}
}
