% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummaryPlots.r
\name{generateSummaryPlots}
\alias{generateSummaryPlots}
\title{Plotting function to make a eGSEA summary plot}
\usage{
generateSummaryPlots(plot.data, file.name = "resumPlot",
  Xlab = "-log10(p-value)", Ylab = "Average Absolute logFC",
  format = NULL, firstN = 10, savePlot = TRUE, legend = TRUE)
}
\arguments{
\item{plot.data}{A dataframe formatted for this function. Output of coGSEA$resumPlot2}

\item{file.name}{Name of  the  file to save. Not needed if savePlot = FALSE (character)}

\item{Xlab}{label for X axis (character)}

\item{firstN}{N number of top N genes to highlight in blue (integer)}

\item{savePlot}{Wheter to save the plot, or just display it (boolean)}

\item{Ybal}{label for Y axis (character)}

\item{legen}{whether to display or not the legend if savePlot = FALSE (boolean)}
}
\value{
A plot if savePlot = TRUE, else nothing
}
\description{
Plotting function to make a eGSEA summary plot
}
\examples{
\dontrun{
generateSummaryPlots(plot.data = coGSEAres$resumPlot2, file.name = "resumPlot", Xlab="-log10(p-value)",Ylab="Average Absolute logFC", format = "pdf", firstN = 10,  savePlot = TRUE)
}
}
